Add method to create snapshot creator in test base
